Fixed groupId name
