+++ pom.xml
@@ -8,7 +8,7 @@
 		<version>3.1.0</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>com.luizalabs</groupId>
+	<groupId>com.dev</groupId>
 	<artifactId>wishlist-api</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>wishlist-api</name>
