+++ src/main/java/com/dev/wishlist/WishlistApplication.java
@@ -1,4 +1,4 @@
-package com.luizalabs.wishlist;
+package com.dev.wishlist;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
