+++ src/main/java/com/dev/wishlist/mappers/ExceptionResponseMapper.java
@@ -1,7 +1,7 @@
-package com.luizalabs.wishlist.mappers;
+package com.dev.wishlist.mappers;
 
-import com.luizalabs.wishlist.exceptions.ExceptionResponse;
-import com.luizalabs.wishlist.exceptions.GlobalException;
+import com.dev.wishlist.exceptions.ExceptionResponse;
+import com.dev.wishlist.exceptions.GlobalException;
 import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
