+++ src/main/java/com/dev/wishlist/repositories/WishlistRepository.java
@@ -1,6 +1,6 @@
-package com.luizalabs.wishlist.repositories;
+package com.dev.wishlist.repositories;
 
-import com.luizalabs.wishlist.models.Wishlist;
+import com.dev.wishlist.models.Wishlist;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
