+++ src/main/java/com/dev/wishlist/utils/APIConstants.java
@@ -1,4 +1,4 @@
-package com.luizalabs.wishlist.utils;
+package com.dev.wishlist.utils;
 
 public class APIConstants {
