+++ src/test/java/com/dev/wishlist/TestWishlistApplication.java
@@ -1,4 +1,4 @@
-package com.luizalabs.wishlist;
+package com.dev.wishlist;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.test.context.TestConfiguration;
