+++ src/test/java/com/dev/wishlist/WishlistApplicationTests.java
@@ -1,4 +1,4 @@
-package com.luizalabs.wishlist;
+package com.dev.wishlist;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
